+++ genium/src/main/java/com/here/genium/model/cpp/CppStruct.kt
@@ -28,6 +28,7 @@ class CppStruct(
     isExternal: Boolean = false,
     val fields: List<CppField> = emptyList(),
     val methods: List<CppMethod> = emptyList(),
+    val constants: List<CppConstant> = emptyList(),
     val isEquatable: Boolean = false,
     val isImmutable: Boolean = false
 ) : CppExternableElement(name, fullyQualifiedName, comment, isExternal) {